// src/app/shared/types/enums.ts
export enum TripMode {
  Return = 'Return',
  OneWay = 'One-way',
}

export enum BookingClass {
  Economy = 'Economy',
  PremiumEconomy = 'Premium Economy',
  BusinessClass = 'Business Class',
  FirstClass = 'First Class',
}

export enum VehicleType {
  Aircraft = 'Aircraft',
  Train = 'Train',
  Bus = 'Bus'
}

export enum Sort {
  Quality = 'Quality',
  Price = 'Price',
  Duration = 'Duration',
  Date = 'Date',
}